[FEATURE] Add t3o_membership member data synchronization
A new field "related_member" is added to the agency table that
can contan a uid of a t3o_membership member record.

If such a record is present, 3 things happen:

1. All fields from the member table are displayed as read-only
   in the Backend
2. When a agency record is changed in the Backend a data handler
   hook synchronizes the agency data with the member data.
3. When the t3o_membership import task is run, the agency data
   is also updated with the member data.

Additionally a new field "endtime" is added to agencies. This
allows the synchronization of the "endtime" field of the members
that is set when a member has cancelled his subscription.